requirements.txt
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
google-api-core==0.1.2
google-api-python-client==1.6.4
google-auth==1.2.1
google-cloud-core==0.28.0
google-cloud-vision==0.29.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.8.2
httplib2==0.10.3
idna==2.6
oauth2client==4.1.2
olefile==0.44
Pillow==4.3.0
protobuf==3.5.0.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytz==2017.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
